Extract logic into useIndexSet hook and move useIndexDefaults hook
