+++ graylog2-web-interface/src/components/indices/hooks/useIndexDefaults.ts
